requirement.txt
pytorch
torchvision
numpy
random
copy
opencv-python
scipy
nltk
h5py
json
tqdm
collections
PIL
re
lpips
scikit-image
rouge
matplotlib
math
argparse